// ui.frontend/src/almLib/models/ChannelModels.ts
/**
Copyright 2026 Adobe. All rights reserved.
This file is licensed to you under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License. You may obtain a copy
of the License at http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software distributed under
the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
OF ANY KIND, either express or implied. See the License for the specific language
governing permissions and limitations under the License.
*/
export interface EVCChannel {
  id: string;
  name: string;
  description: string;
  boardId: number | null;
  isSubscribed: boolean;
  enabled: boolean;
  state: string;
  recordingCount: number;
  paletteIndex: number;
  colorBase: string;
  newCount: number;
  deletedCount: number;
  recordings: EVCRecording[];
  newRecordings: EVCRecording[];
  deletedRecordings: EVCRecording[];
  isPseudo?: boolean;
}

export interface EVCRecording {
  id: string;
  title: string;
  speaker: string;
  date: string;
  tag: string;
  duration: string;
  url: string;
  sourceKey: string;
  postId: string;
  recordingId: string | null;
  status: 'ACTIVE' | 'DELETED';
  isDeleted: boolean;
  isNew: boolean;
  durationSeconds: number | null;
  publishedDateMs: number | null;
  createdDateMs: number | null;
  activityType: string;
  channelId: number;
  channelName: string;
  boardId: number | null;
  thumbUrl: string;
  color: string;
  titleBg: string;
  titleText: string;
  dateTxt: string;
  badgeColor: string;
  speakerInitials: string;
}